feat: atifacts seggregation
