+++ src/common/interfaces/EnvironmentDataInterface.ts
