+++ src/renderer/utils/defaultUpdateFrequencies.ts
