+++ src/renderer/utils/globalContainerVariants.ts
